+++ src/cloudmesh/ai/command/vm/cloud.py
@@ -1,4 +1,4 @@
-from .providers import cmd as providers_cmd
+from .provider import cmd as providers_cmd
 
 # Alias 'cloud' to 'providers'
 cmd = providers_cmd